+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 5.0.2.0 (beta) - 2025-02-XX
 * Fixed Telescopius CSV format change.
+* Recheck project minimum time after a meridian window clip
 
 ## 5.0.1.0 (beta) - 2025-02-11
 * Added profile preference to enable/disable slew/center for new targets.